to forbid add ,revmove ,sub_qty,add_qty <--->any uesr to any user

+++ orders/views.py
@@ -51,21 +51,24 @@ def cart(request):
 def removecart(request,id):
     if request.user.is_authenticated and not request.user.is_anonymous and id:
         orderdetails = OrderDetails.objects.get(id=id)
-        orderdetails.delete()
+        if orderdetails.order.user.id==request.user.id:
+            orderdetails.delete()
     return redirect('cart')
 
 def add_qty(request, id):
     if request.user.is_authenticated and not request.user.is_anonymous and id:
         orderdetails = OrderDetails.objects.get(id=id)
-        orderdetails.quantity += 1  # corrected the syntax here
-        orderdetails.save()
+        if orderdetails.order.user.id==request.user.id:
+            orderdetails.quantity += 1  # corrected the syntax here
+            orderdetails.save()
     return redirect('cart')
 
 
 def sub_qty(request, id):
     if request.user.is_authenticated and not request.user.is_anonymous and id:
         orderdetails = OrderDetails.objects.get(id=id)
-        if orderdetails.quantity > 1:
-            orderdetails.quantity -= 1
-            orderdetails.save()
+        if orderdetails.order.user.id==request.user.id:
+            if orderdetails.quantity > 1:
+                orderdetails.quantity -= 1
+                orderdetails.save()
     return redirect('cart')
